修复重量取值的问题

+++ middle-trade/src/main/java/com/pousheng/middle/order/dto/ShipmentExtra.java
@@ -92,5 +92,5 @@ public class ShipmentExtra implements Serializable{
     private String mposShipmentId;
 
     //重量
-    private Long weight;
+    private double weight;
 }
