+++ middle-web/src/main/java/com/pousheng/middle/open/api/dto/YyEdiShipInfo.java
@@ -33,5 +33,5 @@ public class YyEdiShipInfo {
     /**
      * 发货重量
      */
-    private Long weight;
+    private double weight;
 }
